requirements.txt
Flask 
Flask-SQLAlchemy
marshmallow
redis
requests
celery
pytest